feat: Apply flexbox layout to arrange the main sections
The three main containers within the body flex container
are the header, content, and footer. The header and
footer are rows with minimal item positioning. The
main containers for the content section are the hero,
information, quote, and welcome sections. Hero,
quote, and welcome are row directed, while the
information section is column directed. Quote and
welcome have nested containers that are row directed.