feat: Modify the color palette to differ from the initial colors
The initial color palette was dark navy blue for the
header, hero, and footer. The quote section was gray,
and the information and welcome sections were white.
The style sheet now styles the header, hero, and footer
to be dark gray, the information and welcome sections
are light gray, and the quote section is a darker light
gray than the body. The buttons and image borders are
colored dark green.